support shared state between server instances
Add a shared state between all FTP servers running in the system.
It is based on a shared memory with concurrent access protected
by a semaphore.

ftok() is used to generate semaphore and shared meory keys based
on argv[0]. Make sure all vsfpd instances are started with the same
path to the binary.