count system wide sessions
Config option max_clients applies only to sessions spawned
by the current FTP server. Add counting semaphore as a
resource shared by all servers initialized with max_clients
of the first FTP server that becomes active.

This way we can start one FTP server on IPv4  and another
one with similar configuration but on IPv6 and limit the
total number of FTP connections accepted by the system.

SEM_UNDO is used together with decreasing the semaphore value
so that the counter is automatically incremented by the system
in case the session crashes.